runtime: Integrate content-directory with working group.
