+++ runtime/src/integration/mod.rs
@@ -1,4 +1,3 @@
-pub mod content_directory;
 pub mod proposals;
 pub mod storage;
 pub mod transactions;
